+++ core/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `MessageOps::value_mut()`, which returns a mutable reference to a message's value, if any.
+- Add `SerializedValue::convert()`, `SerializedValueSlice::convert()` and
+  `MessageOps::convert_value()`.
 
 ### Changed